Version 1.2 - Added logic so if the script is deployed as SYSTEM it will create a scheduled task to run the script for the current logged on user
If the Toast script is deployed in the SYSTEM context, the script source is copied to a new folder in the users %TEMP% Directory.

The folder is given a unique GUID name.

A scheduled task is created for the current logged on user and is unique for the each time the Toast Script is deployed. Each scheduled task is named using the User SID and the unique Task GUID.

If the script is deployed to the current logged on user, a scheduled task is not created and the script is run as normal